// graphql/__generated__/book-ql.types.d.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
};
export type AddRes = {
    __typename?: 'AddRes';
    message?: Maybe<Scalars['String']>;
    status?: Maybe<Scalars['String']>;
};
export type Book = {
    __typename?: 'Book';
    author?: Maybe<Scalars['String']>;
    genre?: Maybe<Scalars['String']>;
    id?: Maybe<Scalars['ID']>;
    title?: Maybe<Scalars['String']>;
};
export type BookFilterInput = {
    author?: InputMaybe<Scalars['String']>;
    genre?: InputMaybe<Scalars['String']>;
    title?: InputMaybe<Scalars['String']>;
};
export type BookUpdateInput = {
    author?: InputMaybe<Scalars['String']>;
    genre?: InputMaybe<Scalars['String']>;
    id: Scalars['ID'];
    title?: InputMaybe<Scalars['String']>;
};
// ...
export type UpdateRes = {
    __typename?: 'UpdateRes';
    book?: Maybe<Book>;
    message?: Maybe<Scalars['String']>;
    status?: Maybe<Scalars['String']>;
};